fix: extract to teamUpdater command
