+++ server/routes/api/team.ts
@@ -1,6 +1,7 @@
 import Router from "koa-router";
+import teamUpdater from "@server/commands/teamUpdater";
 import auth from "@server/middlewares/authentication";
-import { Event, Team } from "@server/models";
+import { Team } from "@server/models";
 import { authorize } from "@server/policies";
 import { presentTeam, presentPolicies } from "@server/presenters";
 import { assertUuid } from "@server/validation";
@@ -19,66 +20,35 @@ router.post("team.update", auth(), async (ctx) => {
     defaultCollectionId,
     defaultUserRole,
   } = ctx.body;
+
   const { user } = ctx.state;
   const team = await Team.findByPk(user.teamId);
   authorize(user, "update", team);
 
-  if (subdomain !== undefined && process.env.SUBDOMAINS_ENABLED === "true") {
-    team.subdomain = subdomain === "" ? null : subdomain;
+  if (defaultCollectionId !== undefined && defaultCollectionId !== null) {
+    assertUuid(defaultCollectionId, "defaultCollectionId must be uuid");
   }
 
-  if (name) {
-    team.name = name;
-  }
-  if (sharing !== undefined) {
-    team.sharing = sharing;
-  }
-  if (documentEmbeds !== undefined) {
-    team.documentEmbeds = documentEmbeds;
-  }
-  if (guestSignin !== undefined) {
-    team.guestSignin = guestSignin;
-  }
-  if (avatarUrl !== undefined) {
-    team.avatarUrl = avatarUrl;
-  }
-
-  if (defaultCollectionId !== undefined) {
-    if (defaultCollectionId !== null) {
-      assertUuid(defaultCollectionId, "defaultCollectionId must be uuid");
-    }
-    team.defaultCollectionId = defaultCollectionId;
-  }
-
-  if (collaborativeEditing !== undefined) {
-    team.collaborativeEditing = collaborativeEditing;
-  }
-
-  if (defaultUserRole !== undefined) {
-    team.defaultUserRole = defaultUserRole;
-  }
-
-  const changes = team.changed();
-  const data = {};
-  await team.save();
-
-  if (changes) {
-    for (const change of changes) {
-      data[change] = team[change];
-    }
-
-    await Event.create({
-      name: "teams.update",
-      actorId: user.id,
-      teamId: user.teamId,
-      data,
-      ip: ctx.request.ip,
-    });
-  }
+  const updatedTeam = await teamUpdater({
+    params: {
+      name,
+      avatarUrl,
+      subdomain,
+      sharing,
+      guestSignin,
+      documentEmbeds,
+      collaborativeEditing,
+      defaultCollectionId,
+      defaultUserRole,
+    },
+    user,
+    team,
+    ip: ctx.request.ip,
+  });
 
   ctx.body = {
-    data: presentTeam(team),
-    policies: presentPolicies(user, [team]),
+    data: presentTeam(updatedTeam),
+    policies: presentPolicies(user, [updatedTeam]),
   };
 });
